Factor out test file specifically for ExprBuilder
