+++ packages/core/query/QueryPlanNode.test.ts
@@ -399,80 +399,6 @@ Deno.test("QueryBuilder LEFT JOINS", async () => {
   >()
 })
 
-Deno.test("QueryBuilder .in()", async () => {
-  const { db } = await init()
-  const plan = dbSchema.query()
-    .from("cats")
-    .where((t) => t.column("cats", "name").in("fluffy", "Mr. Blue"))
-    .select({ name: (t) => t.column("cats", "name") })
-    .plan()
-  expect(plan.describe()).toEqual(
-    `Select(name AS name, Filter(TableScan(default.cats), In(name, ["fluffy", "Mr. Blue"]))) AS $0`,
-  )
-  expect(await plan.execute(db).toArray()).toEqual([
-    { "$0": { name: "fluffy" } },
-    { "$0": { name: "Mr. Blue" } },
-  ])
-})
-
-Deno.test("QueryBuilder .in() with expressions", async () => {
-  const dbSchema = s.db().withTables(
-    s.table("names").with(
-      s.column("name", s.type.string()),
-      s.column("nickname", s.type.string()),
-    ),
-  )
-  const db = await PaulDB.inMemory()
-  const model = await db.dbFile.getDBModel(dbSchema)
-  await model.names.insertMany([
-    { name: "Alice", nickname: "Ali" },
-    { name: "Bob", nickname: "Bob" },
-    { name: "Charlie", nickname: "Chuck" },
-  ])
-
-  const rows = await db.query(
-    dbSchema.query()
-      .from("names")
-      .where((t) =>
-        // this is a contrived example, but it shows that you can use expressions
-        // in the `in` clause. This finds all rows where the `name` is equal to
-        // the `nickname`, or where the `name` is equal to "Chuck"
-        t.column("names", "nickname").in("Chuck", t.column("names.name"))
-      )
-      .select({
-        name: (t) => t.column("names.name"),
-        nickname: (t) => t.column("names.nickname"),
-      }),
-  ).toArray()
-  expect(rows).toEqual([
-    { name: "Bob", nickname: "Bob" },
-    { name: "Charlie", nickname: "Chuck" },
-  ])
-})
-
-Deno.test("QueryBuilder .not()", async () => {
-  const { db } = await init()
-  const plan = dbSchema.query()
-    .from("cats")
-    .where((t) => t.column("cats", "name").eq("fluffy").not())
-    .select({ name: (t) => t.column("cats", "name") })
-    .plan()
-  expect(plan.describe()).toEqual(
-    `Select(name AS name, Filter(TableScan(default.cats), NOT(Compare(name = "fluffy")))) AS $0`,
-  )
-  expect(await plan.execute(db).toArray()).toEqual([
-    { "$0": { name: "mittens" } },
-    { "$0": { name: "Mr. Blue" } },
-  ])
-
-  const plan2 = dbSchema.query()
-    .from("cats")
-    .where((t) => t.not(t.column("cats", "name").eq("fluffy")))
-    .select({ name: (t) => t.column("cats", "name") })
-    .plan()
-  expect(plan2.describe()).toEqual(plan.describe())
-})
-
 Deno.test("QueryBuilder .aggregate()", async () => {
   const { db } = await init()
   const plan = dbSchema.query()
